moving functions to app data utils
